frontend: refactor user and game autocomplete
