+++ frontend/cognitive-app-frontend/src/app/common/game-search/game-search.component.html
